+++ core-adapter-ktx/src/main/java/com/nullgr/core/adapter/ktx/BaseViewHolder.kt
@@ -1,0 +1,12 @@
+package com.nullgr.core.adapter.ktx
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+
+/**
+ * Base ViewHolder for all [AdapterDelegate] that supports view caching.
+ *
+ * @author vchernyshov
+ */
+open class BaseViewHolder(parent: ViewGroup, layoutId: Int)
+    : ViewHolder(LayoutInflater.from(parent.context).inflate(layoutId, parent, false))